% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add_last_datapoint_for_potential_peak.R
\name{Add_last_datapoint_for_potential_peak}
\alias{Add_last_datapoint_for_potential_peak}
\title{Forced additional peak detection}
\usage{
Add_last_datapoint_for_potential_peak(
  last_datapoint,
  last_local_max,
  last_local_min,
  peaks
)
}
\arguments{
\item{last_datapoint}{Last point of curve}

\item{last_local_max}{Last locl max found}

\item{last_local_min}{Last local min found}

\item{peaks}{Hash object with peak information}
}
\value{
Updated peak information
}
\description{
Defines whether or not force_additional_peak is TRUE or FALSE in automatic peak detection
}
\examples{
see vignettes
}
