% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Average_neighbouring_points_filtering.R
\name{Average_neighbouring_points_filtering}
\alias{Average_neighbouring_points_filtering}
\title{Peaks filtering}
\usage{
Average_neighbouring_points_filtering(
  peaks,
  smoothing_data,
  neighbour_temp_window_size
)
}
\arguments{
\item{peaks}{Peaks information}

\item{smoothing_data}{Smoothed data, output of TL_curve_smoothing()}

\item{neighbour_temp_window_size}{Window size}
}
\value{
Updated peak information
}
\description{
Discards peaks that aren't significantly higher than neighbouring points
}
\examples{
see vignettes
}
