% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_parameters.R
\name{compare_parameters}
\alias{compare_parameters}
\title{Comparison of parameters}
\usage{
compare_parameters(param_set1, param_set2)
}
\arguments{
\item{param_set1}{First set of parameters (same format as RLumModel::.set_pars())}

\item{param_set2}{Second set of parameters (same format as RLumModel::.set_pars())}
}
\value{
Common and non-common parameters printed in console.
}
\description{
Comparison of parameters
}
\examples{
compare_parameters(RLumModel::.set_pars("Bailey2001"), RLumModel::.set_pars("Bailey2001"))
}
