% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Complete_unmixing_routine.R
\name{Complete_unmixing_routine}
\alias{Complete_unmixing_routine}
\title{Complete unmixing routine}
\usage{
Complete_unmixing_routine(
  mgeo,
  sgeo,
  crgeo,
  source_list,
  Population_fraction,
  filename,
  error_threshold = NULL,
  score_threshold = NULL,
  nb_tracer = NULL,
  tracer_pair = NULL
)
}
\arguments{
\item{mgeo}{Data frame containing one of the dataset mixtures}

\item{sgeo}{Data frame containing the sediment sources from a dataset}

\item{crgeo}{Data frame containing the CR score for each tracer (from fingerPro::cr_ns())}

\item{source_list}{List of sources c("S1", "S2",...)}

\item{Population_fraction}{Expected population fraction}

\item{filename}{Name of file}

\item{error_threshold}{Select only tracers that have an error below error_threshold}

\item{score_threshold}{Select only tracers that have a score above score_threshold}

\item{tracer_pair}{Pair extracted from \code{\link[fingerPro:pairs]{fingerPro::pairs()}} output}

\item{nb_tracers}{Number of tracers to be selected (ordered from smallest to largest error)}
}
\value{
Results of unmixing and metrics to be able to evaluate quality of unmixing
}
\description{
Performs the entire unmixing routine for 2 ou 3 sources. Additional sources should be easy to add
}
\examples{
see vignettes
}
