% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_mean_median_proportions_differences.R
\name{Compute_mean_median_proportions_differences}
\alias{Compute_mean_median_proportions_differences}
\title{Computation of the differences between the mean/median fractions and the expected population fractions}
\usage{
Compute_mean_median_proportions_differences(
  source_list,
  Population_fraction,
  proportions,
  filename,
  tracer_pair = NULL
)
}
\arguments{
\item{source_list}{List of sources c("S1", "S2",...)}

\item{Population_fraction}{Expected populations proportions}

\item{proportions}{Information of every source contained in a df: filename    variable    mean    median    sd    n    Q25    Q75    lower.ci    upper.ci    error_threshold    score_threshold    nb_tracer    Expected_population_fraction}

\item{filename}{Name of file}

\item{tracer_pair}{Pair of tracers used}
}
\value{
Differences
}
\description{
If the expected population fraction is available, this function computes the differences between the mean/median and the expected population fractions
}
\examples{
see vignettes
}
