% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_OSL_metrics.R
\name{Compute_OSL_metrics}
\alias{Compute_OSL_metrics}
\title{Compute and store OSL metrics}
\usage{
Compute_OSL_metrics(sequence, data)
}
\arguments{
\item{sequence}{Set sequence to model}

\item{data}{\linkS4class{RLum.Analysis} containing OSL record}
}
\value{
Hash object containing OSL metrics (cross-section, relative cross-section, intensity, sensitivity)
}
\description{
Computes OSL metrics using Get_OSL_metrics() and stores the results in hash object
}
\examples{
see vignettes
}
