% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_TL_metrics.R
\name{Compute_TL_metrics}
\alias{Compute_TL_metrics}
\title{Compute and store TL metrics}
\usage{
Compute_TL_metrics(sequence, data, automatic_peak_finder)
}
\arguments{
\item{sequence}{Set sequence to model}

\item{data}{\linkS4class{RLum.Analysis} of TL records}

\item{automatic_peak_finder}{Define if the TL peak detection should be done manually (FALSE) or automatically (TRUE).
Automatic peak detection performs well on synthetically generated TL curves. For measured TL curves, verification is needed to make sure peaks were detected correctly.
If not, set automatic_peak_finder = FALSE and select TL peaks manually.}
}
\value{
Hash object containing TL metrics
}
\description{
Computes and stores TL metrics of all TL records of sequence
}
\examples{
see vignettes
}
