% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consecutive_min_max_intensity_threshold_check.R
\name{Consecutive_min_max_intensity_threshold_check}
\alias{Consecutive_min_max_intensity_threshold_check}
\title{Clean extrema found automatically}
\usage{
Consecutive_min_max_intensity_threshold_check(
  local_extrema,
  peaks,
  min_max_intensity_threshold = 30
)
}
\arguments{
\item{local_extrema}{Data frame containing local extrema}

\item{peaks}{Peak information}

\item{min_max_intensity_threshold}{Distance between extremas}
}
\value{
Updated peak information
}
\description{
Attempt to clean extremas found automatically by deleting extremax that are too close to each other in
order to not fit noise buy actual extremas
}
\examples{
see vignettes
}
