% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_sequence_manually.R
\name{Create_sequence_manually}
\alias{Create_sequence_manually}
\title{Create sequence manually}
\usage{
Create_sequence_manually()
}
\value{
Sequence with the same format as output of function \code{\link[RLumModel:read_SEQ2R]{RLumModel::read_SEQ2R()}}
}
\description{
Create any sequence containing IRR, PH, TL and OSL steps manually. Additional steps can be easy added to this function.
}
\examples{
see vignettes
}
