% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIY_TL_peak_intensities_integration.R
\name{DIY_TL_peak_intensities_integration}
\alias{DIY_TL_peak_intensities_integration}
\title{Manual integration of peak intensities}
\usage{
DIY_TL_peak_intensities_integration(TL_tgcd_results)
}
\arguments{
\item{TL_tgcd_results}{Output of \code{\link[tgcd:tgcd]{tgcd::tgcd()}}}
}
\value{
Integrated peak intensities
}
\description{
Manual integration of peak intensities
}
\examples{
see vignettes
}
