% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_OSL_TL_metrics_for_all_records.R
\name{Extract_OSL_TL_metrics_for_all_records}
\alias{Extract_OSL_TL_metrics_for_all_records}
\title{Extract TL/OSL metrics}
\usage{
Extract_OSL_TL_metrics_for_all_records(
  TL_hash,
  OSL_hash,
  source_number,
  records_to_extract_data_from = NULL
)
}
\arguments{
\item{TL_hash}{Hash object containing TL metrics}

\item{OSL_hash}{Hash object containing OSL metrics}

\item{source_number}{Source number ("S1" or "S2" or ...)}

\item{records_to_extract_data_from}{Records from which to extract data (c("Record_1", "Record_2")).}
}
\value{
OSL/TL metrics for all records
}
\description{
Extract TL/OSL metrics
}
\examples{
see vignettes
}
