% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_peaks_using_std.R
\name{Filter_peaks_using_std}
\alias{Filter_peaks_using_std}
\title{Filter peaks using standard deviation}
\usage{
Filter_peaks_using_std(
  local_extrema,
  peaks,
  df,
  std_computation_window_size_degrees,
  std_threshold
)
}
\arguments{
\item{local_extrema}{All local extrema}

\item{peaks}{Peak information}

\item{df}{Curve data}

\item{std_computation_window_size_degrees}{Window size in °C}

\item{std_threshold}{Standard deviation threshold}
}
\value{
Updated peak information
}
\description{
Filters peaks based on intensity std of window around peak.
Keeps only peaks that have std of intensity around them greater than std_threshold
}
\examples{
see vignettes
}
