% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_synthetic_OSL_TL_signals.R
\name{Generate_synthetic_OSL_TL_signals}
\alias{Generate_synthetic_OSL_TL_signals}
\title{Exact copy of \code{\link[RLumModel:model_LuminescenceSignals]{RLumModel::model_LuminescenceSignals()}}}
\usage{
Generate_synthetic_OSL_TL_signals(sequence, model = "Bailey2001")
}
\arguments{
\item{sequence}{Set sequence to model}

\item{model}{\link{character} (\strong{required}): set model to be used. Available models are:
\code{"Bailey2001"}, \code{"Bailey2002"}, \code{"Bailey2004"}, \code{"Pagonis2007"}, \code{"Pagonis2008"}, \code{"Friedrich2017"}, \code{"Friedrich2018"}, \verb{"Peng2022}" and for own models \code{"customized"} (or \code{"customised"}).
Note: When model = \code{"customized"}/\verb{"customised} is set, the argument \code{own_parameters} has to be set.}
}
\value{
This function returns an \link[Luminescence:RLum.Analysis-class]{Luminescence::RLum.Analysis} object with all TL, (LM-) OSL and RF/RL steps
in the sequence. Every entry is an \link[Luminescence:RLum.Data.Curve-class]{Luminescence::RLum.Data.Curve} object and can be plotted, analysed etc. with
further \code{RLum}-functions.
}
\description{
Initially contained more code after model_LuminescenceSignals() and used that function in various scripts. Ended only keeping
model_LuminescenceSignals() in that function but kept it in order to not have to modify already written scripts. For thourough
documentation, refer to \code{\link[RLumModel:model_LuminescenceSignals]{RLumModel::model_LuminescenceSignals()}}.
}
\examples{
see vignettes
}
