% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_initial_peaks.R
\name{Get_initial_peaks}
\alias{Get_initial_peaks}
\title{Automatic peak detection (local max)}
\usage{
Get_initial_peaks(df, span)
}
\arguments{
\item{df}{Data frame containing TL intensity and temperature}

\item{span}{Span for function inflect()}
}
\value{
Peak information
}
\description{
Automatic peak detection (local max)
}
\examples{
see vignettes
}
