% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_initial_peaks_valleys.R
\name{Get_initial_peaks_valleys}
\alias{Get_initial_peaks_valleys}
\title{Local extrema}
\usage{
Get_initial_peaks_valleys(df, span)
}
\arguments{
\item{df}{Data frame containing TL intensity and temperature}

\item{span}{Span for function inflect()}
}
\value{
Local extrema
}
\description{
Extract local extrema (min and max)
}
\examples{
see vignettes
}
