% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IRR_user_inputs.R
\name{get_IRR_user_inputs}
\alias{get_IRR_user_inputs}
\title{Get irradiation (IRR) step info}
\usage{
get_IRR_user_inputs()
}
\value{
IRR temperature (°C) / dose (Gy) / dose rate (Gy/s) / time (s)
}
\description{
Get IRR temperature (°C) / dose (Gy) / dose rate (Gy/s) / time (s)
}
\examples{
see vignettes

}
