% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_metrics_for_proprotions_selection.R
\name{Get_metrics_for_proprotions_selection}
\alias{Get_metrics_for_proprotions_selection}
\title{Extract metrics from estimated proportions}
\usage{
Get_metrics_for_proprotions_selection(
  metrics,
  filename,
  tracer_pair,
  ctsgeo,
  source_list,
  ratio_differences,
  proportions,
  error_threshold,
  score_threshold,
  nb_tracer
)
}
\arguments{
\item{metrics}{Data frame contining unmixing metrics}

\item{filename}{Name of file}

\item{tracer_pair}{Pair extracted from \code{\link[fingerPro:pairs]{fingerPro::pairs()}} output}

\item{ctsgeo}{Data frame containing scores and errors for the best tracers}

\item{source_list}{List of sources c("S1", "S2",...)}

\item{ratio_differences}{Mean/median differences between estimated and expected population fractions}

\item{proportions}{Information of every source contained in a df: filename    variable    mean    median    sd    n    Q25    Q75    lower.ci    upper.ci    error_threshold    score_threshold    nb_tracer    Expected_population_fraction}

\item{error_threshold}{Select only tracers that have an error below error_threshold}

\item{score_threshold}{Select only tracers that have a score above score_threshold}

\item{nb_tracers}{Number of tracers to be selected (ordered from smallest to largest error)}
}
\value{
Updated metrics data frame
}
\description{
This function updated data frame "metrics" with information from unmixing routine
such as the mean/median Manhattan distance (if expected population fraction provided),
number of tracers/score/error thresholds used to select tracers for unmixing, 95\% CI, width of CI
}
\examples{
see vignettes
}
