% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_OSL_metrics.R
\name{Get_OSL_metrics}
\alias{Get_OSL_metrics}
\title{Extract OSL metrics}
\usage{
Get_OSL_metrics(model.output, IRR_dose, make_plot = TRUE)
}
\arguments{
\item{model.output}{\linkS4class{RLum.Analysis} for a single record}

\item{IRR_dose}{Irradiation dose}

\item{make_plot}{Logical. Define whether or not a plot should be made}
}
\value{
Number of OSL components, photoionisation cross section, intensity/sensitivity of each component
}
\description{
Extract OSL metrics
}
\examples{
see vignettes
}
