% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OSL_user_inputs.R
\name{get_OSL_user_inputs}
\alias{get_OSL_user_inputs}
\title{Get OSL info}
\usage{
get_OSL_user_inputs()
}
\value{
OSL temperature (°C) / durations (s) / optical power (\%)
}
\description{
Get OSL temperature (°C) / durations (s) / optical power (\%)
}
\examples{
see vignettes
}
