% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_param_normal_dist_list_of_values.R
\name{Get_param_normal_dist_list_of_values}
\alias{Get_param_normal_dist_list_of_values}
\title{Prepare normal distribution of parameter for every depth value for sandbox rule book}
\usage{
Get_param_normal_dist_list_of_values(
  parameter_name,
  populations_parameters,
  n_depth
)
}
\arguments{
\item{parameter_name}{Name of parameter to act upon}

\item{populations_parameters}{Parameters of quartz luminescence models for every population}

\item{n_depth}{Number of depth values}
}
\value{
list of list with normal distribution for 'parameter_name'
}
\description{
In sandbox rule book, at every depth step, the value of the parameters must be defined. It is often done by using a normal distribution.
Note that in function \code{\link[=Get_param_normal_dist_list_of_values]{Get_param_normal_dist_list_of_values()}}, the sd = 0. The function basically does the same as (for parameter N1 and two populations):

par_N1 <- list(list(mean = seq(from = 1.5e+07, to = 1.5e+07, length.out = n_depth),
sd = seq(from = 0, to = 0, length.out = n_depth)),# Default N1 for pop1
list(mean = seq(from = 0, to = 0, length.out = n_depth),
sd = seq(from = 0, to = 0, length.out = n_depth)))# No N1 for pop2
}
\examples{
see vignettes
}
