% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_user_inputs.R
\name{get_peak_user_inputs}
\alias{get_peak_user_inputs}
\title{Number of peaks user input}
\usage{
get_peak_user_inputs(record_num)
}
\arguments{
\item{record_num}{}
}
\value{
Number of peaks and whether there is an incomplete (forced_additional_peak = TRUE) peak at the end
}
\description{
Query user to define total number of peaks.
Forced_additional_peak = TRUE means that at the end of the curve, there seems to be the beginning of an incomplete peak.
Needs to be selected (and forced_additional_peak = TRUE) in order to not disrupt the fitting of the previous curves. No metrics
are extracted from this incomplete peak
}
\examples{
see vignettes
}
