% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_peaks.R
\name{Get_peaks}
\alias{Get_peaks}
\title{Peak detection}
\usage{
Get_peaks(df, span)
}
\arguments{
\item{df}{Data containing TL intensity and temperature}

\item{span}{Number of points around the peak to consider}
}
\value{
Peaks number and location
}
\description{
Automatic peak detection
Span is the number of points around the peak to consider
https://www.rdocumentation.org/packages/ggpmisc/versions/0.6.0/topics/find_peaks
}
\examples{
see vignettes
}
