% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PH_user_inputs.R
\name{get_PH_user_inputs}
\alias{get_PH_user_inputs}
\title{Get Preheat (PH) step info}
\usage{
get_PH_user_inputs()
}
\value{
PH temperature (째C) / duration (s) / heating rate (째C/s)
}
\description{
Get PH temperature (째C) / duration (s) / heating rate (째C/s)
}
\examples{
see vignettes
}
