% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_specific_columns.R
\name{Get_specific_columns}
\alias{Get_specific_columns}
\title{Get specific column by (partial-) name}
\usage{
Get_specific_columns(fit, col_prefix, components)
}
\arguments{
\item{fit}{Output of function \code{\link[Luminescence:fit_CWCurve]{Luminescence::fit_CWCurve()}}}

\item{col_prefix}{Prefix that should be contained in the desired columns}

\item{components}{Component number}
}
\value{
Subset containing only the desired columns
}
\description{
Selects columns that start with a specific prefix and contain a specific component number
}
\examples{
see vignettes
}
