% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_TL_metrics.R
\name{Get_TL_metrics}
\alias{Get_TL_metrics}
\title{Routine to obtain TL metrics}
\usage{
Get_TL_metrics(
  record_num,
  model.output,
  Heat_rate,
  span,
  automatic_peak_finder,
  INITIAL_ACTIVATION_ENERGY = 1.4,
  INITIAL_KINETIC_ORDER = 1.5
)
}
\arguments{
\item{record_num}{Record number ("Record_x")}

\item{model.output}{Single TL record of \linkS4class{RLum.Analysis}}

\item{Heat_rate}{Heat rate}

\item{span}{Span for function \code{\link[=Get_initial_peaks_valleys]{Get_initial_peaks_valleys()}}}

\item{automatic_peak_finder}{Define if the TL peak detection should be done manually (FALSE) or automatically (TRUE).
Automatic peak detection performs well on synthetically generated TL curves. For measured TL curves, verification is needed to make sure peaks were detected correctly.
If not, set automatic_peak_finder = FALSE and select TL peaks manually.}

\item{INITIAL_ACTIVATION_ENERGY}{Activation energy}

\item{INITIAL_KINETIC_ORDER}{Kinetic order}
}
\value{
TL peaks metrics
}
\description{
Routine to obtain TL metrics
}
\examples{
see vignettes
}
