% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TL_user_inputs.R
\name{get_TL_user_inputs}
\alias{get_TL_user_inputs}
\title{Get TL info}
\usage{
get_TL_user_inputs()
}
\value{
TL begin/end temperature (째C) and heat rate (째C/s)
}
\description{
Get TL begin/end temperature (째C) and heat rate (째C/s)
}
\examples{
see vignettes

}
