% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_number_of_sequence_steps_input.R
\name{get_user_number_of_sequence_steps_input}
\alias{get_user_number_of_sequence_steps_input}
\title{Get number of sequence steps from user}
\usage{
get_user_number_of_sequence_steps_input()
}
\value{
Number of steps in the manually created sequence
}
\description{
Get user input on the number of step in sequence when creating it manually
}
\examples{
see vignettes
}
