% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_sequence_step_input.R
\name{get_user_sequence_step_input}
\alias{get_user_sequence_step_input}
\title{Define sequence steop}
\usage{
get_user_sequence_step_input(i)
}
\arguments{
\item{i}{Step number in the sequence}
}
\value{
Sequence step PH, IRR, OSL or TL
}
\description{
Define sequence steop
}
\examples{
see vignettes
}
