% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflect.R
\name{inflect}
\alias{inflect}
\title{Find local extrema}
\usage{
inflect(x, threshold = 1)
}
\arguments{
\item{x}{Data points}

\item{threshold}{Window size}
}
\value{
Local extrema
}
\description{
Find local extrema based on \url{https://stackoverflow.com/questions/6836409/finding-local-maxima-and-minima}
}
\examples{
see vignettes
}
