% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keep_relevant_peaks.R
\name{Keep_relevant_peaks}
\alias{Keep_relevant_peaks}
\title{Filter peaks out of all the automatically selected peaks}
\usage{
Keep_relevant_peaks(
  local_extrema,
  df,
  smoothing_data,
  min_max_intensity_threshold = 20,
  std_computation_window_size_degrees = 75,
  std_threshold = 0
)
}
\arguments{
\item{local_extrema}{All local extrema}

\item{df}{Curve data}

\item{smoothing_data}{Smoothed curve data}

\item{min_max_intensity_threshold}{Threshold for \code{\link[=Consecutive_min_max_intensity_threshold_check]{Consecutive_min_max_intensity_threshold_check()}}}

\item{std_computation_window_size_degrees}{Window size for \code{\link[=Average_neighbouring_points_filtering]{Average_neighbouring_points_filtering()}}}

\item{std_threshold}{Threshold for \code{\link[=Filter_peaks_using_std]{Filter_peaks_using_std()}}}
}
\value{
Updated peaks
}
\description{
Filter peaks out of all the automatically selected peaks
}
\examples{
see vignettes
}
