% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_custom_model_and_retrieve_OSL_TL_metrics.R
\name{Make_custom_model_and_retrieve_OSL_TL_metrics}
\alias{Make_custom_model_and_retrieve_OSL_TL_metrics}
\title{Extract metrics of modeled luminescence signals}
\usage{
Make_custom_model_and_retrieve_OSL_TL_metrics(
  sequence,
  dose_rate,
  own_parameters,
  own_state_parameters
)
}
\arguments{
\item{sequence}{\link{list} (\strong{required}): set sequence to model as \link{list} or as *.seq file from the
Risø sequence editor.}

\item{dose_rate}{\link{numeric} (\strong{required}): laboratory dose rate in XXX
Gy/s for calculating seconds into Gray in the *.seq file.}

\item{own_parameters}{\link{list} (\strong{required}): This argument allows the user to submit own parameter sets. See details for more information.}

\item{own_state_parameters}{\link{numeric} (\strong{required}): Some publications (e.g., Pagonis 2009)
offer state parameters. With this argument the user can submit this state parameters. For further details
see vignette "RLumModel - Using own parameter sets" and example 3. The parameter also accepts an
\link[Luminescence:RLum.Results-class]{Luminescence::RLum.Results} object created by \link{.set_pars} as input.}
}
\value{
OSL and TL metrics
}
\description{
Extracts metrics from OSL and TL curves generated by \code{\link[RLumModel:model_LuminescenceSignals]{RLumModel::model_LuminescenceSignals()}}
}
\examples{
see vignettes
}
