% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_luminescence_rulebook.R
\name{Make_luminescence_rulebook}
\alias{Make_luminescence_rulebook}
\title{Create Luminescence RuleBook}
\usage{
Make_luminescence_rulebook(populations_parameters, pop_ratios)
}
\arguments{
\item{populations_parameters}{Bailey2001 parameters for every population in the RuleBook}

\item{pop_ratios}{Ratios of populations in the mix}
}
\value{
Rulebook with luminescence model properties as rules
}
\description{
Sets sandbox RuleBook with luminescence properties as rules + strict minimum properties mandatory in the rule book (grainsize distribution, age).
Rules should be changed/added according to specific needs.
}
\examples{
see vignettes
}
