% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_SAR_OSL_copy.R
\name{measure_SAR_OSL_copy}
\alias{measure_SAR_OSL_copy}
\title{Slightly modified copy of sandbox::measure_SAR_OSL()}
\usage{
measure_SAR_OSL_copy(aliquot, sequence, dose_rate = 0.1)
}
\arguments{
\item{aliquot}{\link{data.frame} or a \link{list} of it, a set of grains that are assigned to an
aliquot (sample subset used for measurement), i.e., the result of
\link{prepare_Aliquot}.}

\item{sequence}{\link{list}, definition of the SAR protocol.}

\item{dose_rate}{\link{numeric} value, Dose rate of the luminescence
reader, in Gy/s.}
}
\value{
\link[Luminescence:RLum.Analysis-class]{Luminescence::RLum.Analysis} object. Equivalent of the import result for
a real world measurement file. This object can be evaluated by functions
of the package \code{Luminescence-package}.
}
\description{
The function models the time-dependent photon counts of an aliquot
according to the specified CW SAR OSL (continuous wave, single aliquot
regenerative dose protocol for optically stimulated luminescence) sequence
and parameters. The modelling is done for each component and photon count
curves are summed to return an \link[Luminescence:RLum.Analysis-class]{Luminescence::RLum.Analysis} object as equivalent of
importing a real measurement data set to the R-package \code{Luminescence-package}.

The function uses the package \link[RLumModel:RLumModel-package]{RLumModel::RLumModel-package} to perform the simulation of the
photon count curves.

It was modified to fit the need of RLumUnmix package. Modifications were made clear in code.
}
\examples{
see vignettes
}
\author{
Michael Dietze, GFZ Potsdam (Germany),
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom),
Modified by Balthazar Allegri
}
