% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSL_metrics_hash_to_dataframe.R
\name{OSL_metrics_hash_to_dataframe}
\alias{OSL_metrics_hash_to_dataframe}
\title{Convert hash object containing OSL metrics to data frame}
\usage{
OSL_metrics_hash_to_dataframe(OSL_hash, record_num)
}
\arguments{
\item{OSL_hash}{Hash object containing OSL metrics}

\item{record_num}{Record number ("Record_x")}
}
\value{
Data frame containing OSL metrics
}
\description{
Convert hash object containing OSL metrics to data frame
}
\examples{
see vignettes
}
