% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_smoothing.R
\name{Plot_smoothing}
\alias{Plot_smoothing}
\title{Plot TL curve smoothing}
\usage{
Plot_smoothing(df, smoothing_results, record_num)
}
\arguments{
\item{df}{Curve data}

\item{smoothing_results}{Smoothing data}

\item{record_num}{Record number ("Record_x")}
}
\value{
Plot of smoothed TL curve above un-smoothed TL curve
}
\description{
Plot TL curve smoothing
}
\examples{
see vignettes
}
