% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_sources_proportions.R
\name{Plot_sources_proportions}
\alias{Plot_sources_proportions}
\title{Boxplot of estimated population proportions + true proportions if provided}
\usage{
Plot_sources_proportions(
  result_FP_long,
  filename,
  data_sol,
  Population_fraction,
  source_list,
  tracer_pair = NULL
)
}
\arguments{
\item{result_FP_long}{Result of unmixing}

\item{filename}{Name of file}

\item{data_sol}{Data frame containing tracers used for unmixing. Output of function \code{\link[=Select_data_for_unmixing]{Select_data_for_unmixing()}}}

\item{Population_fraction}{Expected population fraction}

\item{source_list}{List of sources c("S1", "S2",...)}

\item{tracer_pair}{Pair of tracers used}
}
\value{
Boxplot with estimated and expected population fractions
}
\description{
Boxplot of estimated population proportions + true proportions if provided
}
\examples{
see vignettes
}
