% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_TL_Curve.R
\name{Plot_TL_Curve}
\alias{Plot_TL_Curve}
\title{TL curve plotting}
\usage{
Plot_TL_Curve(model_output, peaks, record_num)
}
\arguments{
\item{model_output}{TL curve data}

\item{peaks}{Local peaks}

\item{record_num}{Record number ("Record_x")}
}
\value{
Plut of the TL curve
}
\description{
TL curve plotting
}
\examples{
see vignettes
}
