% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_TL_Curve_with_extrema.R
\name{Plot_TL_Curve_with_extrema}
\alias{Plot_TL_Curve_with_extrema}
\title{Plot TL curve with extrema}
\usage{
Plot_TL_Curve_with_extrema(model_output, local_extrema, record_num)
}
\arguments{
\item{model_output}{TL curve data}

\item{local_extrema}{All local extrema}

\item{record_num}{Record number ("Record_x")}
}
\value{
Plot of TL curve with extrema
}
\description{
Plot TL curve with extrema
}
\examples{
see vignettes
}
