% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_TL_Curve_with_std.R
\name{Plot_TL_Curve_with_std}
\alias{Plot_TL_Curve_with_std}
\title{Plot normalised intensity standard deviation}
\usage{
Plot_TL_Curve_with_std(model_output, temp_std_results, record_num)
}
\arguments{
\item{model_output}{TL curve data}

\item{temp_std_results}{Standard deviation data}

\item{record_num}{Record number ("Record_x")}
}
\value{
Plot of normalised intensity standard deviation
}
\description{
Plot normalised intensity standard deviation
}
\examples{
see vignettes
}
