% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_TL_Peaks.R
\name{Plot_TL_Peaks}
\alias{Plot_TL_Peaks}
\title{Plot fitted peaks individually}
\usage{
Plot_TL_Peaks(df, npeak)
}
\arguments{
\item{df}{Output of \code{\link[tgcd:tgcd]{tgcd::tgcd()}}}

\item{npeak}{Total number of fitted peaks}
}
\value{
Plot of fitted peaks
}
\description{
Plot peaks fitted with \code{\link[tgcd:tgcd]{tgcd::tgcd()}} individually
}
\examples{
see vignettes
}
