% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_parameters_from_aliquot.R
\name{prep_parameters_from_aliquot}
\alias{prep_parameters_from_aliquot}
\title{First part of measure_SAR_OSL_copy()}
\usage{
prep_parameters_from_aliquot(aliquot)
}
\arguments{
\item{aliquot}{\link{data.frame} or a \link{list} of it, a set of grains that are assigned to an
aliquot (sample subset used for measurement), i.e., the result of
\link{prepare_Aliquot}.}
}
\value{
parameters
}
\description{
First part of measure_SAR_OSL_copy()
}
\examples{
see vignettes
}
