% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_data_for_tracer_selection.R
\name{Prepare_data_for_tracer_selection}
\alias{Prepare_data_for_tracer_selection}
\title{Data preparation for unmixing}
\usage{
Prepare_data_for_tracer_selection(data_path)
}
\arguments{
\item{data_path}{Path to the dataframe containing data of sources and mixes}
}
\value{
Data needed for the unmixing routine as well as the original "stacked_df" loaded.
}
\description{
The df stored in data_path should have the same structure as the df named 'unmixing_df' (in the data folder)
First column should be named "sources": each source in that column should be named  "S1", "S2", "S3", ... and "Mixed"
The following columns should be the potential tracers
The last column should be named "Pop_fraction" and should contain the population fraction of each source. Not mandatory
}
\examples{
see vignettes
}
