% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rename_sequence.R
\name{Rename_sequence}
\alias{Rename_sequence}
\title{Rename sequence}
\usage{
Rename_sequence(list_of_lists)
}
\arguments{
\item{list_of_lists}{Sequence}
}
\value{
Renamed sequence with unique names
}
\description{
Rename sequence
}
\examples{
se vignettes
}
