% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_data_for_unmixing.R
\name{Select_data_for_unmixing}
\alias{Select_data_for_unmixing}
\title{Select data used for unmixing}
\usage{
Select_data_for_unmixing(sgeo, mgeo, unmixing_tracers)
}
\arguments{
\item{sgeo}{Data frame containing the sediment sources from a dataset}

\item{mgeo}{Data frame containing one of the dataset mixtures}

\item{unmixing_tracers}{Tracers selected for unmixing}
}
\value{
Data frame containing all the information needed for the unmixing
}
\description{
Select data used for unmixing
}
\examples{
see vignettes
}
