% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_unmixing_tracers_2_sources.R
\name{Select_unmixing_tracers_2_sources}
\alias{Select_unmixing_tracers_2_sources}
\title{Selection of tracers for unmixing of 2 sources}
\usage{
Select_unmixing_tracers_2_sources(
  mgeo,
  sgeo,
  crgeo,
  score_threshold = NULL,
  nb_tracers = NULL
)
}
\arguments{
\item{mgeo}{Data frame containing one of the dataset mixtures}

\item{sgeo}{Data frame containing the sediment sources from a dataset}

\item{crgeo}{Data frame containing the CR score for each tracer (from fingerPro::cr_ns())}

\item{score_threshold}{Select only tracers that have a score above score_threshold}

\item{nb_tracers}{Number of tracers to be selected (ordered from smallest to largest error)}
}
\value{
Data frame contained scored of best tracers selected
}
\description{
Selection of tracers for unmixing of 2 sources
}
\examples{
see vignettes
}
