% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Separate_numeric_and_non_numeric_part.R
\name{Separate_numeric_and_non_numeric_part}
\alias{Separate_numeric_and_non_numeric_part}
\title{Separate numeric and non-numeric part of parameter name}
\usage{
Separate_numeric_and_non_numeric_part(parameter_name)
}
\arguments{
\item{parameter_name}{Name of parameter}
}
\value{
Numeric and non-numeric parts of parameter_name
}
\description{
Separates 'osl_E_th1' into 'E_th' and 1
}
\examples{
Separate_numeric_and_non_numeric_part('osl_E_th1')
}
