% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Rule_for_all_model_parameters.R
\name{set_Rule_for_all_model_parameters}
\alias{set_Rule_for_all_model_parameters}
\title{Sets rule in RuleBook for all model parameters}
\usage{
set_Rule_for_all_model_parameters(book, depth, populations_parameters)
}
\arguments{
\item{book}{Sandbox RuleBook}

\item{depth}{depth \link{numeric} \link{list}, specifying the depths used for the interpolation.}

\item{populations_parameters}{Parameters of quartz luminescence models for every population}
}
\value{
Updated RuleBook
}
\description{
Applies function \code{\link[=Update_parameter_in_rulebook]{Update_parameter_in_rulebook()}} to the first 7 parameters of model 'Bailey2001'
}
\examples{
see vignettes
}
