% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_record_TL_OSL_metrics_df.R
\name{Single_record_TL_OSL_metrics_df}
\alias{Single_record_TL_OSL_metrics_df}
\title{Aggregates TL and OSL metrics for single record}
\usage{
Single_record_TL_OSL_metrics_df(TL_hash, OSL_hash, record_num)
}
\arguments{
\item{TL_hash}{Hash object containing TL metrics}

\item{OSL_hash}{Hash object containing OSL metrics}

\item{record_num}{Record number ("Record_x")}
}
\value{
Single line data frame containing all OSL and TL metrics
}
\description{
Aggregates TL and OSL metrics for single record
}
\examples{
see vignettes
}
