% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL_curve_smoothing.R
\name{TL_curve_smoothing}
\alias{TL_curve_smoothing}
\title{Smoothing TL curve}
\usage{
TL_curve_smoothing(df, record_num, lambda, make_smooth_plot = FALSE)
}
\arguments{
\item{df}{Curve data}

\item{record_num}{Record number ("Record_x")}

\item{lambda}{Smoothing parameter (rough: 50 / smooth: 1e4)}

\item{make_smooth_plot}{}
}
\value{
Smoothed TL curve
}
\description{
Smoothens TL curve using \code{\link[pracma:whittaker]{pracma::whittaker()}}.
}
\examples{
see vignette
}
