% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL_metrics_hash_to_dataframe.R
\name{TL_metrics_hash_to_dataframe}
\alias{TL_metrics_hash_to_dataframe}
\title{Converting hash to Data Frame}
\usage{
TL_metrics_hash_to_dataframe(TL_hash, record_num)
}
\arguments{
\item{TL_hash}{Hash object containing TL metrics}

\item{record_num}{TL record number ("Record_x")}
}
\value{
Data frame
}
\description{
Converts hash object containing TL metris to data frame
}
\examples{
see vignettes
}
